+++ SkillModifiers/time_dilation.lua
@@ -21,8 +21,10 @@ time_dilation:set_add_func(function(data, modifier_index, item_id)
             last_frame = current_frame
             time_dilation_flag = true
             gm.audio_play_sound(sound, 1.0, true)
+            gm.game_set_speed(120, false)
             Alarm.destroy(alarm)
             alarm = Alarm.create(function()
+                gm.game_set_speed(60, false)
                 gm.audio_stop_sound(sound)
                 time_dilation_flag = false
             end, during)
@@ -45,14 +47,11 @@ local black_list = {
     [gm.constants.oBlackOut] = true,
     [gm.constants.oB] = true,
     [gm.constants.oHUD] = true,
-    [gm.constants.oInit] = true
+    [gm.constants.oBossSpawn] = true,
+    [gm.constants.oBossSpawn2] = true,
+    [gm.constants.oRope] = true,
+    [gm.constants.oBNoSpawn] = true
 }
-local need_to_step_list = {
-    [gm.constants.oInit] = true,
-    [gm.constants.oDirectorControl] = true
-}
-local cache_table = {}
-local last_current_frame
 memory.dynamic_hook("event_perform_internal", "int64_t", {"CInstance*", "RValue*", "int", "int", "int"},
     Dynamic.event_perform_internal, function(ret_val, target, result, object_index, event_type, event_number)
         if not time_dilation_flag then
@@ -66,27 +65,8 @@ memory.dynamic_hook("event_perform_internal", "int64_t", {"CInstance*", "RValue*
             return
         end
         if gm.object_is(object_index_, gm.constants.pEnemy) then
-            if gm.variable_global_get("_current_frame") % 2 == 0 then
+            if gm.variable_global_get("_current_frame") % 4 ~= 0 then
                 return false
-            end
-        elseif gm.object_is(object_index_, gm.constants.pFriend) or need_to_step_list[object_index] then
-            local current_frame = gm.variable_global_get("_current_frame")
-            if current_frame % 2 == 0 then
-                if current_frame ~= last_current_frame then
-                    cache_table = {}
-                    last_current_frame = current_frame
-                end
-                local number = event_number:get()
-                local target_cache = cache_table[target.id]
-                if target_cache == nil then
-                    target_cache = {}
-                    cache_table[target.id] = target_cache
-                end
-                if not target_cache[number] then
-                    target_cache[number] = true
-                    target:event_perform(3, number)
-                    target:event_perform(3, number)
-                end
             end
         end
     end)